modify sql and finish the lottery logic

+++ vazee.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `users` (
 `id` INT(11) NOT NULL AUTO_INCREMENT COMMENT 'id',
-`phone` INT(11) NOT NULL COMMENT '手机号码',
+`phone` BIGINT(20) NOT NULL COMMENT '手机号码',
 `user_name` VARCHAR(32) NOT NULL COMMENT '用户名字',
 `ctime` INT(11) COMMENT '操作时间',
 PRIMARY KEY (`id`, `phone`)
